app/requirements.txt
fastmcp>=2.8.1
paramiko>=3.5.1
pysocks>=1.7.1